requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
mysql-connector-python
python-dotenv
httpx